leitura do ficheiro projects (incompleto) tarefas a andar

+++ src/Projeto/Design.java
@@ -7,7 +7,7 @@ public class Design extends Task {
     public Design() {
     }
 
-    public Design(String name,Date startDate, Date endDate, int duration, int conclusionState, Person responsible, float effortRate) {
+    public Design(String name,Date startDate, Date endDate, int duration, int conclusionState, Person responsible, double effortRate) {
         super(name, startDate, endDate, duration, conclusionState, responsible, effortRate);
     }
