+++ src/Projeto/Development.java
@@ -7,7 +7,7 @@ public class Development extends  Task {
 
     }
 
-    public Development(String name, Date startDate, Date endDate, int duration, int conclusionState, Person responsible, float effortRate) {
+    public Development(String name, Date startDate, Date endDate, int duration, int conclusionState, Person responsible, double effortRate) {
         super(name, startDate, endDate, duration, conclusionState, responsible, effortRate);
     }
